the issue of Keychain is found: x509 V3 is not supported (on Sierra at least)
The only problem was setting the version of ssl certificate to 3. After
commenting out the line 94 from root_ca_generate_from_key.c, Keychain
recognized the certificate output file. However, since this is
Keychain's fault, we will not deal with it.

Also, I removed the extension of NID_key_usage:"critical,keyCertSign,cRLSign"
as it seems an unnecessary constraint at least for the root certificate.
We don't know the answer yet, therefore we leave it be.