use only public key for creating the root certificate, only engine knows the private
It's not possible to test it for now and it seems irrational if we defer
testing when the engine is ported to the rust part.
It would be better if we write a stub openssl program which takes the
private key and signs a single message digest. Anyway, if the tests
pass, it seems we are all done with root_ca_generate_from_key.c file